+++ tests/test_app/webroot/js/test2.js
@@ -1,0 +1,4 @@
+var first = 'This is first';
+var second = 'This is second';
+
+alert(first + ' and ' + second);